backend/requirements.txt
# Core Framework
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
python-multipart>=0.0.9

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1

# File Processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Database
motor==3.3.1
pymongo==4.5.0

# LLM Providers
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
langchain>=0.1.0

# Security & Auth
python-jose>=3.3.0
pyjwt>=2.10.1
passlib>=1.7.4
cryptography>=42.0.8

# HTTP & Requests
requests>=2.31.0
requests-oauthlib>=2.0.0
httpx>=0.25.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Utilities
python-multipart>=0.0.9
email-validator>=2.2.0
tzdata>=2024.2
typer>=0.9.0

# AWS (if needed)
boto3>=1.34.129

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Monitoring (Optional)
# sentry-sdk[fastapi]>=1.40.0
# prometheus-fastapi-instrumentator>=6.1.0